# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dify Python Code Validator",
    "version": "1.0.0",
    "description": "检查 Python 3.12 代码并在一次性隔离容器中运行 Agent 提供的 pytest 断言。 API 无需 Bearer API Key。"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "请替换为 Dify 可访问的 API_PUBLIC_URL。"
    }
  ],
  "paths": {
    "/v1/python/check": {
      "post": {
        "operationId": "python_static_check",
        "summary": "检查 Python 源码",
        "description": "检查语法、Ruff 静态诊断以及导入模块是否存在于固定 Python 3.12 执行镜像中。不会执行提交的源码。若有错误，请修复后再次调用。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "静态检查结果。status 为 passed 表示没有语法、lint 或导入错误。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/python/test": {
      "post": {
        "operationId": "python_assertion_test",
        "summary": "在沙箱中测试 Python 源码",
        "description": "将 source 写入 solution.py，将 test_code 写入 pytest 测试文件，在无网络、资源受限的一次性容器中运行。至少需要一个通过的测试才返回 passed；测试代码由 Agent 提供。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "pytest 执行结果。运行超时、资源超限、失败、错误和未收集到测试均有独立状态。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResponse"
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CodeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "description": "完整的单个 Python 3.12 源码文件内容。UTF-8 输入总大小上限为 256 KiB。"
          }
        }
      },
      "TestRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "source",
          "test_code"
        ],
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "description": "将被测试的 Python 源码。"
          },
          "test_code": {
            "type": "string",
            "minLength": 1,
            "description": "pytest 测试源码，至少包含一个 test_* 测试。可使用 from solution import ...。"
          }
        }
      },
      "Diagnostic": {
        "type": "object",
        "required": [
          "source",
          "code",
          "message",
          "severity"
        ],
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "syntax",
              "ruff",
              "imports",
              "checker"
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "error",
              "warning"
            ]
          },
          "line": {
            "type": "integer",
            "nullable": true
          },
          "column": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "CheckResponse": {
        "type": "object",
        "required": [
          "status",
          "syntax_valid",
          "python_version",
          "diagnostics",
          "checked_imports"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "error"
            ]
          },
          "syntax_valid": {
            "type": "boolean"
          },
          "python_version": {
            "type": "string"
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            }
          },
          "checked_imports": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TestResponse": {
        "type": "object",
        "required": [
          "status",
          "passed",
          "failed",
          "errors",
          "skipped",
          "duration_ms",
          "stdout",
          "stderr",
          "output_truncated"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "passed",
              "failed",
              "error",
              "timed_out",
              "resource_limited",
              "no_tests"
            ]
          },
          "passed": {
            "type": "integer",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "minimum": 0
          },
          "errors": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "minimum": 0
          },
          "exit_code": {
            "type": "integer",
            "nullable": true
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0
          },
          "stdout": {
            "type": "string"
          },
          "stderr": {
            "type": "string"
          },
          "output_truncated": {
            "type": "boolean"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "RequestTooLarge": {
        "description": "请求体超过 1 MiB 传输限制。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "请求字段缺失、格式无效或源码超过 256 KiB 限制。"
      },
      "TooManyRequests": {
        "description": "并发执行槽位已满，请稍后重试。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unavailable": {
        "description": "执行器不可用或沙箱镜像尚未构建。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}